refactor: :truck: move page files to own directories
