+++ src/pages/projects/index.tsx
@@ -1,10 +1,10 @@
 import { useState, useEffect } from 'react'
 import { Grid, Title } from '@mantine/core'
-import { projects } from '../utils/projects'
-import Layout from '../components/Layout/Layout'
-import Project from '../components/Project/Project'
-import CustomLoader from '../components/CustomLoader/CustomLoader'
-import { IProjectProps } from '../interfaces/Project.interface'
+import { projects } from '../../utils/projects'
+import Layout from '../../components/Layout/Layout'
+import Project from '../../components/Project/Project'
+import CustomLoader from '../../components/CustomLoader/CustomLoader'
+import { IProjectProps } from '../../interfaces/Project.interface'
 
 const Projects = () => {
   const [isLoading, setIsLoading] = useState<boolean>(true)
@@ -21,6 +21,7 @@ const Projects = () => {
       <Title order={1} mb={30}>
         Projects
       </Title>
+
       <Grid gutter="xl">
         {isLoading ? (
           <CustomLoader />
